+++ pydobe/__init__.py
@@ -1,2 +1,3 @@
 from pydobe.after_effects.objects.root import Root
+
 objects = Root()  # entry point to the root level ExtendScript objects available