+++ pydobe/adobe_objects.py
@@ -1,7 +1,8 @@
 from pydobe.core import PydobeBaseObject
 
+
 class File(PydobeBaseObject):
-    def __init__(self,  pydobe_id=None):
+    def __init__(self, pydobe_id=None):
         super().__init__(pydobe_id)
 
     def __str__(self):
@@ -13,16 +14,16 @@ class File(PydobeBaseObject):
 
     @property
     def full_name(self) -> str:
-        return self._eval_on_this_object('fullName')
+        return self._eval_on_this_object("fullName")
 
     "The file name portion of the absolute URI, without the path specification."
 
     @property
     def name(self) -> str:
-        return self._eval_on_this_object('name')
+        return self._eval_on_this_object("name")
 
     "The path portion of the absolute URI, without the file name"
 
     @property
     def path(self) -> str:
-        return self._eval_on_this_object('path')
+        return self._eval_on_this_object("path")