+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2023-02-09
+
+### Added
+
+- Following functionality added:
+  - Add footage to project
+  - Replace footage in project
 
 ## [0.2.0] - 2023-02-09
 
@@ -47,3 +54,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.0.1]: https://github.com/LisaGG89/pydobe/releases/tag/v0.0.1
 [0.1.0]: https://github.com/LisaGG89/pydobe/releases/tag/v0.1.0
 [0.2.0]: https://github.com/LisaGG89/pydobe/releases/tag/v0.2.0
+[0.3.0]: https://github.com/LisaGG89/pydobe/releases/tag/v0.3.0